NVDLA -- make NVDLA configurable on nv_small and nv_large.
nv_small is passing four flatbuf tests.
nv_large failed on compilation due to missing sram IP.

+++ actions/hdl_nvdla/hw/action_config.sh
@@ -21,8 +21,12 @@
 
 . $SNAP_ROOT/snap_env.sh
 
-# NVDLA specific variables
-NVDLA_CONFIG=nv_large
+echo "NVDLA CONFIG: ${NVDLA_CONFIG}"
+
+if [ -z ${NVDLA_CONFIG} ]; then
+    echo "NVDLA CONFIG is empty, please specify it in snap_env.sh"
+    exit -1
+fi
 
 if [ -L ./nvdla ]; then
     unlink ./nvdla
@@ -56,7 +60,11 @@ if [ ! -d ../nvdla-capi ]; then
     echo "WARNING!!! Please use 'git submodule init' to initialize nvdla hardware IP."
     exit -1
 elif [ ! -d ../nvdla-capi/outdir/$NVDLA_CONFIG ]; then
-    cd ../nvdla-capi/; make USE_NV_ENV=1 NV_PROJ=$NVDLA_CONFIG
+    cd ../nvdla-capi/
+    if [ -f tree.make ]; then
+        rm tree.make
+    fi
+    make USE_NV_ENV=1 NV_PROJ=$NVDLA_CONFIG
     ./tools/bin/tmake -clean -build vmod
     if [ $? -ne 0 ]; then
         echo "ERROR while making NVDLA-CAPI hardware."
@@ -74,6 +82,11 @@ ln -s $ACTION_ROOT/nvdla-capi/outdir/$NVDLA_CONFIG/vmod/fifos               fifo
 ln -s $ACTION_ROOT/nvdla-capi/outdir/$NVDLA_CONFIG/vmod/fpga_ip/ram_wrapper ram_wrapper
 ln -s $ACTION_ROOT/nvdla-capi/outdir/$NVDLA_CONFIG/vmod/fpga_ip             ../fpga_ip
 ln -s $ACTION_ROOT/nvdla-capi/outdir/$NVDLA_CONFIG/spec/defs                defs
+
+if [ ! -f defs/project.vh ]; then
+    echo "Cannot find defs/project.vh"
+    exit -1;
+fi
 
 for vsource in *.v_source; do
     vfile=`echo $vsource | sed 's/v_source$/v/'`
@@ -97,6 +110,16 @@ for vsource in *.v_source; do
         sed -i '/#ifdef NVDLA_DBB_ADDR_WIDTH < 64/,/#endif/d' $vfile
     else
         sed -i '/#ifdef NVDLA_DBB_ADDR_WIDTH < 64/d' $vfile
+    fi
+
+    if [ $dbb_data_width -eq 64 ]; then
+        sed -i '/#ifdef NVDLA_DBB_DATA_WIDTH == 256/,/#endif/d' $vfile
+        sed -i '/#ifdef NVDLA_DBB_DATA_WIDTH == 64/d' $vfile
+    fi
+
+    if [ $dbb_data_width -eq 256 ]; then
+        sed -i '/#ifdef NVDLA_DBB_DATA_WIDTH == 64/,/#endif/d' $vfile
+        sed -i '/#ifdef NVDLA_DBB_DATA_WIDTH == 256/d' $vfile
     fi
 
     if [ $sram_data_width ]; then
